Fix layout of firebase ui
