+++ frontend/package.json
@@ -18,6 +18,7 @@
     "firebaseui": "^4.8.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-firebaseui": "^5.0.2",
     "react-scripts": "4.0.3",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
